+++ packages/extension-koni-ui/src/stores/types.ts
@@ -7,7 +7,7 @@ import { AddressBookState, AllLogoMap, AssetSetting, BalanceItem, ChainStakingMe
 import { AccountJson, AccountsContext, AuthorizeRequest, MetadataRequest, SigningRequest } from '@subwallet/extension-base/background/types';
 import { _ChainState } from '@subwallet/extension-base/services/chain-service/types';
 import { SWTransactionResult } from '@subwallet/extension-base/services/transaction-service/types';
-import { WalletConnectSessionRequest } from '@subwallet/extension-base/services/wallet-connect-service/types';
+import { WalletConnectNotSupportRequest, WalletConnectSessionRequest } from '@subwallet/extension-base/services/wallet-connect-service/types';
 import { SessionTypes } from '@walletconnect/types';
 
 import { SettingsStruct } from '@polkadot/ui-settings/types';
@@ -99,6 +99,7 @@ export interface RequestState extends ConfirmationsQueue, BaseReduxStore {
   numberOfConfirmations: number;
   transactionRequest: Record<string, SWTransactionResult>;
   connectWCRequest: Record<string, WalletConnectSessionRequest>;
+  notSupportWCRequest: Record<string, WalletConnectNotSupportRequest>;
 }
 
 export interface UpdateConfirmationsQueueRequest extends BaseReduxStore {